refactor(checker): move type-reference helpers out of core
